+++ src/lib/target-generators/index.ts
@@ -1,4 +1,3 @@
-import * as node from "./node";
 import * as webpack from "./webpack";
 
-export {node, webpack};
+export {webpack};